split exception_details and result into smaller and reusable files
