+++ webdriver/tests/bidi/script/evaluate/primitive_values.py
@@ -1,0 +1,16 @@
+import pytest
+from webdriver.bidi.modules.script import ContextTarget
+from . import PRIMITIVE_VALUES
+
+
+@pytest.mark.asyncio
+@PRIMITIVE_VALUES
+async def test_primitive_values(bidi_session, top_context, expression,
+                                expected):
+    result = await bidi_session.script.evaluate(
+        expression=expression,
+        target=ContextTarget(top_context["context"]),
+        await_promise=True,
+    )
+
+    assert result == expected
